Use built-in DigitalOcean image module
The problem with the ISO was that it is installation media, meaning that
the machine wasn't partitioned or anything on boot, and used a temp
filesystem. The built-in DigitalOcean module builds an image that is
pretty much ready-to-go, especially for DO.

In order to get it to properly boot, I needed to have an option to
remove the default bootloader I had in my base config. I think this is
because the DO module includes a different bootloader (grub) by default.